# wasm: reduce number of VM calls for policy evaluation