--- a/wasm/src/context.h
+++ b/wasm/src/context.h
@@ -23,4 +23,6 @@ opa_value *opa_builtin2(int, void *, opa_value *, opa_value *);
 opa_value *opa_builtin3(int, void *, opa_value *, opa_value *, opa_value *);
 opa_value *opa_builtin4(int, void *, opa_value *, opa_value *, opa_value *, opa_value *);
 
+int32_t eval(opa_eval_ctx_t *ctx);
+
 #endif