… internal/compiler/wasm/opa/opa.go
… internal/compiler/wasm/opa/opa.wasm
… rego/testdata/bundle.tar.gz