This adds a new function export to the wasm modules, `opa_eval`: It offers a faster eval functionality that requires less VM calls to evaluate a policy.

Where before we had to do these VM calls to evaluate a policy with input:

```
<opa_heap_ptr_set>
<opa_eval_ctx_new>
<opa_eval_ctx_set_data>
<opa_eval_ctx_set_entrypoint>
<opa_malloc>
<opa_value_parse>
<opa_eval_ctx_set_input>
<eval>
<opa_eval_ctx_get_result>
<opa_value_dump>
```

Now, it'll be **one** call: `opa_eval`, that takes the entrypoint, a pointer to data, a pointer (+ length) of the input JSON, and a heap ptr. It returns the same result pointer you'd gotten before (as end result of the entire call set).

👉 The added function changes our ABI from `1.1` to `1.2`.

✅ Fixes #3146.

--------

#### TODO

- [x] give the export a better name (💭 `opa_eval`...?)
- [x] bump abi minor version
- [x] add integration test for being able to eval ABI 1.1 module
- [x] docs
- [x] compile some benchmarks
- [x] use new method in `make wasm-rego-test`


------
```
$ benchstat main.bench branch.bench
name         old time/op    new time/op    delta
WasmRego-16    84.3µs ± 6%    15.1µs ± 0%  -82.07%  (p=0.008 n=5+5)

name         old alloc/op   new alloc/op   delta
WasmRego-16    6.05kB ± 0%    2.24kB ± 0%  -62.92%  (p=0.008 n=5+5)

name         old allocs/op  new allocs/op  delta
WasmRego-16       266 ± 0%        65 ± 0%  -75.60%  (p=0.008 n=5+5)
```