@srenatus 2021-07-10
Mishandled the heap ptr in vm.go, we'll keep accumulating inputs, only append them instead of overwriting. Will fix.