@tsandall 2021-07-16
Is there a strong reason to keep those files around? I'm inclined to delete
them.

On Thu, Jul 15, 2021 at 5:01 AM Stephan Renatus ***@***.***>
wrote:

> ***@***.**** commented on this pull request.
> ------------------------------
>
> In rego/testdata/Makefile
> <https://github.com/open-policy-agent/opa/pull/3627#discussion_r670396308>
> :
>
> > @@ -0,0 +1,5 @@
>
> +# NOTE(sr): this is the last image emitting wasm modules with ABI 1.1
>
> +IMAGE := openpolicyagent/opa:0.30.1
>
> +
>
> +bundle.tar.gz: t.rego
>
>
> ☝️ @tsandall <https://github.com/tsandall> just realised I've merged this
> without having resolved that thing. I'll do a cleanup PR if you think
> differently about generated files there! 😃
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/open-policy-agent/opa/pull/3627#discussion_r670396308>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAB2KJMZD7SNHK5M5KC65V3TX3EX5ANCNFSM5ACIJHRA>
> .
>


-- 
-Torin